Fixed fish collision, D-pad alignment, and game over screen clipping.
Extended fish collision boxes so that flipping them wouldn't make them glitch through walls.

Changed alignment of up arrow on D-pad to make it more in line with the other buttons.

Set items on the GameOver screen to clip so that they aren't displayed over the edge of our virtual screen.

Slight renaming of variable.