Fixed starting conditions of Gun and GameTimer.
Previously, Gun would start with a default angle and then suddenly jerk to the beginning of its animation on the next frame, and GameTimer would display a time of 0 for the first couple of frames before correcting to 30. Both of these issues have been fixed. Also, I cut a superfluous method from MainMenu.